+++ maker/src/actor_system.rs
@@ -35,6 +35,8 @@ use xtra::Context;
 use xtra::Handler;
 use xtra_libp2p::listener;
 use xtra_libp2p::Endpoint;
+use xtra_libp2p_ping::ping;
+use xtra_libp2p_ping::pong;
 use xtras::supervisor;
 
 const ENDPOINT_CONNECTION_TIMEOUT: Duration = Duration::from_secs(20);
@@ -124,9 +126,10 @@ where
 
         let (endpoint_addr, endpoint_context) = Context::new(None);
 
-        let ping_address = xtra_libp2p_ping::Actor::new(endpoint_addr.clone(), PING_INTERVAL)
+        ping::Actor::new(endpoint_addr.clone(), PING_INTERVAL)
             .create(None)
             .spawn(&mut tasks);
+        let pong_address = pong::Actor::default().create(None).spawn(&mut tasks);
 
         let endpoint = Endpoint::new(
             TokioTcpConfig::new(),
@@ -134,7 +137,7 @@ where
             ENDPOINT_CONNECTION_TIMEOUT,
             [(
                 xtra_libp2p_ping::PROTOCOL_NAME,
-                xtra::message_channel::StrongMessageChannel::clone_channel(&ping_address),
+                xtra::message_channel::StrongMessageChannel::clone_channel(&pong_address),
             )],
         );
